checked deleteAllCommentsAbove setting before delete comments above

+++ src/option/solveroptionwidget.cpp
@@ -410,9 +410,11 @@ void SolverOptionWidget::addOptionFromDefinition(const QModelIndex &index)
 
             switch(msgBox.exec()) {
             case 0: // replace
-                disconnect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
-                deleteCommentsBeforeOption(indices.at(0).row());
-                connect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
+                if (settings && settings->deleteAllCommentsAboveOption() && indices.size()>0) {
+                    disconnect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
+                    deleteCommentsBeforeOption(indices.at(0).row());
+                    connect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
+                }
                 replaceExistingEntry = true;
                 indices = ui->solverOptionTableView->model()->match(ui->solverOptionTableView->model()->index(0, mOptionTableModel->getColumnEntryNumber()),
                                                                     Qt::DisplayRole,
@@ -436,6 +438,8 @@ void SolverOptionWidget::addOptionFromDefinition(const QModelIndex &index)
             switch(msgBox.exec()) {
             case 0: // delete and replace
                 disconnect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
+                if (settings && settings->deleteAllCommentsAboveOption() && indices.size()>0)
+                   deleteCommentsBeforeOption(indices.at(0).row());
                 deleteOption(true);
                 connect(mOptionTableModel, &SolverOptionTableModel::solverOptionItemRemoved, mOptionTableModel, &SolverOptionTableModel::on_removeSolverOptionItem);
                 replaceExistingEntry = true;
